refactor(auth): 重构认证模块的客户端认证机制
- 将AuthProperties中的timeOut字段类型从Long改为Integer
- 删除ApiClientController和ClientDetailsService相关认证逻辑
- 移除ApiDictController和ApiUserController中的Basic Auth认证
- 添加HeadUtils工具类用于获取请求头信息
- 修改OauthApiClient使用统一的请求头获取方式
- 更新OauthApiEndpointEnum添加注解并调整字典API路径
- 删除OauthStartProperties配置类
- 在SysProjectClientService中添加listByServer方法定义

+++ simple-common-auth/simple-common-auth-client/src/main/java/com/simple/common/auth/client/common/properties/AuthProperties.java
@@ -24,7 +24,7 @@ public class AuthProperties {
     private String serverUrl = "http://localhost:8000";
 
     //客户端获取服务端用户信息的超时时间
-    private Long timeOut = 1000 * 10L;
+    private Integer timeOut = 1000 * 10;
 
     //解密是否校验有效时间
     private Boolean decryptCheckValidityPeriod = false;
